platform/backend/pyproject.toml
[project]
name = "vision-ai-platform-backend"
version = "1.0.0"
description = "Vision AI Training Platform - Production Backend"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # Web Framework
    "fastapi>=0.108.0",
    "uvicorn[standard]>=0.25.0",
    "python-multipart>=0.0.6",

    # Database
    "sqlalchemy>=2.0.23",
    "psycopg2-binary>=2.9.9",  # PostgreSQL driver for Railway
    "alembic>=1.13.1",         # Database migrations

    # LLM (Dual Provider Support: OpenAI-compatible + Gemini)
    "openai>=1.0.0",
    "google-generativeai>=0.3.0",
    # Legacy LangChain (kept for potential future use)
    "langchain>=0.1.0",
    "langchain-core>=0.1.7",

    # WebSocket
    "python-socketio>=5.10.0",
    "websockets>=12.0",

    # Workflow Orchestration (Phase 12)
    "temporalio>=1.11.0",

    # Redis (State Management & Pub/Sub)
    "redis[hiredis]>=5.0.1",

    # Utils
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",       # For fetching models from Training Services
    "boto3>=1.34.0",          # S3/R2 storage client
    "kubernetes>=34.0.0",     # Kubernetes Job management for training

    # Monitoring & Metrics
    "prometheus-client>=0.19.0",
    "mlflow>=2.9.2",
    "clearml>=1.14.0",

    # Inference Engine (Platform Inference Endpoint)
    "onnxruntime>=1.16.0",    # ONNX Runtime for model inference
    "pillow>=10.0.0",         # Image preprocessing
    "numpy>=1.24.0",          # Array operations

    # Authentication & Security
    "bcrypt==4.0.1",
    "passlib>=1.7.4",
    "python-jose[cryptography]>=3.3.0",
    "PyJWT>=2.8.0",
    "pydantic[email]",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.25.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.5.0",
    "flake8>=6.1.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.25.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.5.0",
    "flake8>=6.1.0",
]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_ignores = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
